make move representation consistently smaller (i8)
This shrinks a lot of data structures by a factor of 4. It should reduce memory usage and maybe improve performance a bit

+++ engine/src/error.rs
@@ -13,16 +13,16 @@ use flip::unflip::FlipStack;
 pub trait FrontEnd : FatalError {
     fn reset_buffer_display(g_timer:&mut Timer);
     fn display_buffers(g_timer: &mut Timer);
-    fn after_update_best_list_verbose(best_list: &mut [i32; 4]);
-    fn before_update_best_list_verbose(best_list: &mut [i32; 4], move_0: i32, best_list_index: i32, best_list_length: &mut i32);
+    fn after_update_best_list_verbose(best_list: &mut [i8; 4]);
+    fn before_update_best_list_verbose(best_list: &mut [i8; 4], move_0: i8, best_list_index: i32, best_list_length: &mut i32);
     fn end_tree_search_output_some_second_stats(alpha: i32, beta: i32, curr_val: i32, update_pv: i32, move_index: i32, echo: i32);
     fn end_tree_search_some_pv_stats_report(alpha: i32, beta: i32, curr_val: i32);
-    fn end_tree_search_level_0_ponder_0_short_report(move_0: i32, first: i32);
+    fn end_tree_search_level_0_ponder_0_short_report(move_0: i8, first: i32);
     fn end_tree_search_output_some_stats(entry: &HashEntry);
     fn end_tree_search_level_0_ponder_0_report(alpha: i32, beta: i32, result: i32, best_move_: i32);
     fn end_tree_search_level_0_report(alpha: i32, beta: i32);
     fn send_solve_status(empties: i32, _side_to_move: i32, eval_info: &mut EvaluationType,
-                         pv_zero: &mut [i32; 64],
+                         pv_zero: &mut [i8; 64],
                          pv_depth_zero: i32,
                          g_timer: &mut Timer,
                          search_state: &mut SearchState);
@@ -32,11 +32,11 @@ pub trait FrontEnd : FatalError {
     fn end_report_panic_abort(elapsed_time: f64);
     fn end_report_semi_panic_abort(elapsed_time: f64);
     fn end_display_zero_status();
-    fn handle_fatal_pv_error(i: i32, pv_0_depth: i32, pv_0: &[i32; 64]);
+    fn handle_fatal_pv_error(i: i32, pv_0_depth: i32, pv_0: &[i8; 64]);
     fn report_unwanted_book_draw(this_move: i32);
     fn report_in_get_book_move_1(side_to_move: i32, remaining_slack: i32, board_state: &mut BoardState, g_book: &mut Book);
     fn report_in_get_book_move_2(chosen_score: i32, chosen_index: i32, flags: &i32, x: &[CandidateMove; 60], search_state: &SearchState);
-    fn midgame_display_simple_ponder_move(move_0: i32);
+    fn midgame_display_simple_ponder_move(move_0: i8);
     fn midgame_display_initial_ponder_move(alpha: i32, beta: i32);
     fn midgame_display_ponder_move(
         max_depth: i32, alpha: i32, beta: i32,
@@ -53,10 +53,10 @@ pub trait FrontEnd : FatalError {
     fn report_mirror_symetry_error(count: i32, i: i32, first_mirror_offset: i32, first_item: i32, second_item: i32);
 }
 pub trait FatalError {
-    fn invalid_move(curr_move: i32) -> !;
+    fn invalid_move(curr_move: i8) -> !;
     fn unrecognized_character(unrecognized: i8) -> !;
     fn cannot_open_game_file(file_name: &str) -> !;
-    fn invalid_move_in_move_sequence(curr_move: i32) -> !;
+    fn invalid_move_in_move_sequence(curr_move: i8) -> !;
     fn internal_error_in_book_code() -> !;
     fn unexpected_character_in_a_move_string() -> !;
     fn invalid_move_string_provided() -> !;
